+++ projects/packages/publicize/src/class-publicize-base.php
@@ -1619,6 +1619,17 @@ abstract class Publicize_Base {
 	/**
 	 * Check if the social image generator is enabled.
 	 *
+	 * @deprecated 0.24.2 use Automattic\Jetpack\Publicize\Publicize_Base\has_social_image_generator_feature instead.
+	 * @param int $blog_id The blog ID for the current blog.
+	 * @return bool
+	 */
+	public function is_social_image_generator_enabled( $blog_id ) { // phpcs:ignore VariableAnalysis.CodeAnalysis.VariableAnalysis.UnusedVariable
+		return $this->has_social_image_generator_feature();
+	}
+
+	/**
+	 * Check if the social image generator is enabled.
+	 *
 	 * @return bool
 	 */
 	public function has_social_image_generator_feature() {