+++ projects/packages/videopress/src/client/block-editor/blocks/video/save.tsx
@@ -10,6 +10,7 @@ import { getVideoPressUrl } from '../../../lib/url';
 /**
  * Types
  */
+import { isVideoFramePosterEnabled } from './components/poster-panel';
 import type { VideoBlockAttributes } from './types';
 import type React from 'react';
 
@@ -44,19 +45,22 @@ export default function save( { attributes }: videoBlockSaveProps ): React.React
 		posterData,
 	} = attributes;
 
-	const { previewOnHover, previewAtTime, previewLoopDuration } = posterData ?? {};
-
 	const blockProps = useBlockProps.save( {
 		className: classnames( 'wp-block-jetpack-videopress', 'jetpack-videopress-player', {
 			[ `align${ align }` ]: align,
 		} ),
 	} );
 
+	const isPreviewOnHoverEnabled = isVideoFramePosterEnabled();
+
+	const autoplayArg = ! isPreviewOnHoverEnabled ? autoplay : autoplay || posterData.previewOnHover;
+	const mutedArg = ! isPreviewOnHoverEnabled ? muted : muted || posterData.previewOnHover;
+
 	const videoPressUrl = getVideoPressUrl( guid, {
-		autoplay: autoplay || posterData.previewOnHover, // enabled when `previewOnHover` is enabled.
+		autoplay: autoplayArg,
 		controls,
 		loop,
-		muted: muted || posterData.previewOnHover, // enabled when `previewOnHover` is enabled.
+		muted: mutedArg,
 		playsinline,
 		preload,
 		seekbarColor,
@@ -76,20 +80,9 @@ export default function save( { attributes }: videoBlockSaveProps ): React.React
 	return (
 		<figure { ...blockProps } style={ style }>
 			{ videoPressUrl && (
-				<>
-					{ previewOnHover && (
-						<span
-							style={ { display: 'none', visibility: 'hidden', position: 'absolute' } }
-							className="videopress-poh"
-						>
-							<span className="videopress-poh__sp">{ previewAtTime }</span>
-							<span className="videopress-poh__duration">{ previewLoopDuration }</span>
-						</span>
-					) }
-					<div className="jetpack-videopress-player__wrapper">
-						{ `\n${ videoPressUrl }\n` /* URL needs to be on its own line. */ }
-					</div>
-				</>
+				<div className="jetpack-videopress-player__wrapper">
+					{ `\n${ videoPressUrl }\n` /* URL needs to be on its own line. */ }
+				</div>
 			) }
 
 			{ ! RichText.isEmpty( caption ) && (