+++ projects/packages/videopress/src/client/components/timestamp-control/types.ts
@@ -8,6 +8,7 @@ export type DecimalPlacesProp = 1 | 2 | 3;
 export type TimestampInputProps = {
 	disabled?: boolean;
 	value: number;
+	min?: number;
 	max?: number;
 	fineAdjustment?: number;
 	onChange?: ( ms: number ) => void;
@@ -19,5 +20,7 @@ export type TimestampControlProps = TimestampInputProps & {
 	label?: ReactNode;
 	help?: ReactNode;
 	wait?: number;
+	marksEvery?: number;
+	renderTooltip?: boolean;
 	onDebounceChange?: ( ms: number ) => void;
 };