+++ projects/packages/wp-js-data-sync/src/contracts/interface-entry-can-delete.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Automattic\Jetpack\WP_JS_Data_Sync\Contracts;
+
+interface Entry_Can_Delete {
+
+	public function delete();
+
+}
+